Fix calcHeight variable
For some reason, event listener doesn't work properly when loading the screen. If you have a theme that has a preloader, it is not positioned properly. I've noticed in the previous code that you are using variable which is on the scope of the context only. Therefore, I've added an easy fix so that iframe is calculated right away and not on load event listener. I've modified calcHeight variable into a global variable to fix the issue.